Replace site with desktop version of new design
The old version of the site gave almost no introduction to me as a
person: it was pretty and very spare, to a fault. This version, in
contrast, lets a visitor know who I am, what I look like, and points
them to my resume and several highlighted projectsin addition to the old
standby links. And links to the twitter account I use for coding shit.

Also gives the code a good scrubbing behind the ears.
* Adds a fluid grid (pocketgrid)
* Content!
  - An about section
  - Lists skills
  - Code examples
  - A nice picture
* Refactor JS
* Total CSS rewrite, using BEM naming conventions