Fix dumb build error with tags
tbh, I think tags are kind of boring and shitty as a UI/navigation
thing. And there were hella errors with the `tagname` variable, which a
quick google didn't resolve.

So rather than going down a dumb and frustrating rabbit hole to fix a
feature I don't even want right now, I just changed the `tag` page so it
just reads "you're it". Fuck it.