Use filenames, not frontmatter, for draft status
Frontmatter gets parsed by both browsers (when hydrating the static
bundle) and node (when building the static bundle), and it is annoyingly
impossible to find a clean, simple condition based on a `draft: boolean`
frontmatter field that will run without error in both environments. This
is a little more annoying to work with, but by god it works.